P2-S1 (absolute-correctness I2): differential oracle finds + fixes 3 real codegen miscompiles
First slice of P2 (behavioral equivalence vs native rustc). Stood up a differential oracle (native
rustc nightly-2026-06-17 vs CARGO_DOTNET_BACKEND=native cargo-dotnet, byte-for-byte) and took the TRUE
current failure census (the fatal type-gate from P1 cannot catch these -- they are behavioral, not
type-safety). Three real codegen bugs fixed, each with a permanent known-answer regression crate
verified byte-identical to native:

1. cilly/src/ir/class.rs -- System.Double/System.Single were emitted as IL `class`, not `valuetype`
   (they ARE .NET value types, like Int128/UInt128). Result: TypeLoadException on EVERY f64/f32
   intrinsic call (min/max/mul_add/powi/MinNumber/MaxNumber/FusedMultiplyAdd/Pow and the rounding
   behind {:.N}). This also DISPROVED the long-suspected flt2dec ULP divergence -- it was this crash
   masquerading; once Double/Single are valuetype, float formatting incl. shortest-repr is
   byte-identical. (cilly/src/ir/tpe/float.rs: f64::is_nan also called Single::IsNaN -> Double::IsNaN.)
   Regression: cargo_tests/float_class_methods.

2. src/terminator/intrinsics/ints.rs -- u128/i128 leading_zeros/trailing_zeros narrowed the 128-bit
   struct result with a raw conv.u4 (invalid on Int128/UInt128) -> garbage (1u128.leading_zeros()
   returned 2386363928, not 127). Now routes the narrow through casts::int_to_int (op_Explicit),
   mirroring the already-correct ctpop arm. Regression: cargo_tests/wideint_ctlz.

3. cilly/src/ir/builtins/atomics.rs -- emulate_subword_cmp_xchng + emulate_subword_xchng declared the
   containing-word-address local (passed as int32& to Interlocked.CompareExchange) as int32 instead of
   int32* -> InvalidProgramException (StackUnexpected). This crashed ALL AtomicU8/I8/U16/I16/Bool
   compare_exchange+swap AND, transitively, std::panic::catch_unwind -- the residual sub-word-atomic
   hazard that has lurked all campaign. Fixed local 0 -> Type::Ptr(i32). Regression:
   cargo_tests/cd_subword_atomics (incl. the catch_unwind-transitive case).

CENSUS (BROKEN_TESTS.md updated to true state, +146): many historical entries proven STALE
(repro clean now -- IPv6/try_fold/i128-saturating/leading_trailing_ones/option-as-slice/hasher/refcell/
ManuallyDrop/rotate/ilog). FUNDAMENTAL walls classified honestly (NOT faked): C-variadic printf
(fuzz47/86/87/96 -- the IL exporter has no vararg calling convention; computation is correct, only the
variadic display diverges; needs a real vararg marshalling shim), half-crate f16/bf16 (CI already
--skip f16). flt2dec ULP reclassified as DISPROVEN. Panic-message stderr-vs-stdout routing classified
REAL-but-orthogonal (deferred to a panic-fidelity workflow; program logic is correct).

VERIFIED: all 3 regression crates re-run on the backend = all checks passed, exit 0 (independently, by
the implement agent, the reviewer, and the parent). Docker ::stable gate GREEN under the FATAL
typechecker: 428 passed / 12 failed = baseline, zero regressions, zero fatal aborts; the backend
compiled in-container, proving the fixes are portable (not native-only). NO checker relaxation
(typecheck.rs unchanged) -- these are codegen corrections, the right resolution; they did not mask a
checker error.

HONEST SCOPE: this is ONE slice of the iterative P2. The full rust-lang/rust core/alloc/std suites were
NOT run differentially through the backend in this env; I2 (exhaustive behavioral equivalence) is NOT
claimed complete. The 3 regression crates are standalone oracle-driven (not yet wired into the
test_lib!/cargo_test! harness) -- a follow-up should register them in CI.

+++ cilly/src/ir/tpe/float.rs
@@ -69,7 +69,9 @@ impl Float {
                 asm.alloc_node(CILNode::call(mref, [val]))
             }
             Float::F64 => {
-                let double = ClassRef::single(asm);
+                // `f64::is_nan` must call `System.Double::IsNaN`, NOT `System.Single::IsNaN`
+                // (copy-paste slip). Calling the f32 method on an f64 arg is a type mismatch.
+                let double = ClassRef::double(asm);
                 let mref = asm.alloc_methodref(MethodRef::new(
                     double,
                     is_nan,
